Move context to contexts dir
